requirements.txt
RPi.GPIO>=0.6.3,<0.7.0
connexion>=1.5.2,<2.0.0
